# apps/messaging-api/openapi-definition.yml
openapi: 3.0.3
info:
  title: OGCIO MessagingIE API
  description: API for OGCIO MessagingIE Service
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/templates/:
    get:
      operationId: ListTemplates
      tags: &a3
        - Templates
      description: Returns the providers matching the requested query
      parameters:
        - schema:
            type: string
            pattern: ^[0-9][0-9]*|undefined$
            default: "0"
          in: query
          name: offset
          required: false
          description: Indicates where to start fetching data or how many records to skip,
            defining the initial position within the list
        - schema:
            type: string
            default: "20"
            pattern: ^([1-9]|100)|undefined$
          in: query
          name: limit
          required: false
          description: Indicates the maximum number (100) of items that will be returned
            in a single request
        - schema:
            type: string
          in: query
          name: search
          required: false
          description: If set, filters by template name in all available languages
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - contents
                      properties:
                        id:
                          type: string
                          format: uuid
                          description: Unique id of the template
                        contents:
                          type: array
                          items:
                            type: object
                            required:
                              - language
                              - templateName
                            properties:
                              language:
                                type: string
                                enum:
                                  - en
                                  - ga
                                description: Template content language
                              templateName:
                                type: string
                                description: Template name for the related language
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        required:
                          - self
                          - first
                          - last
                          - pages
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the request itself
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the next page of results in a paginated response.
                                  If there are no more results, this field may
                                  be omitted
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the previous page of results in a paginated
                                  response. If there are no more results, this
                                  field may be omitted
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          pages:
                            type: object
                            description: It may contain a list of other useful URLs, e.g. one entry for
                              page:'page 1', 'page 2'
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        description: Object containing the links to the related endpoints
                      totalCount:
                        type: number
                        description: Number representing the total number of available items
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
    post:
      operationId: CreateTemplate
      tags: *a3
      description: Creates a new template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contents
              properties:
                contents:
                  type: array
                  items:
                    type: object
                    required:
                      - templateName
                      - language
                      - subject
                      - plainText
                    properties:
                      templateName:
                        type: string
                        description: Template name for the related language
                        minLength: 1
                      language:
                        type: string
                        enum:
                          - en
                          - ga
                        description: Template content language
                      subject:
                        type: string
                        description: Subject of the template
                        minLength: 1
                      excerpt:
                        type: string
                        description: Brief description of the template content
                      plainText:
                        type: string
                        description: Plain text version of the template
                        minLength: 1
                      richText:
                        type: string
                        description: Rich text version of the template
                  minItems: 1
                variables:
                  type: array
                  items:
                    type: object
                    required:
                      - name
                    properties:
                      name:
                        type: string
                        minLength: 1
                  minItems: 1
                  description: List of the variables that are needed to be filled to create a
                    message using this template
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - id
                    properties:
                      id:
                        type: string
                        format: uuid
                  metadata:
                    type: object
                    properties:
                      links:
                        type: object
                        required:
                          - self
                          - first
                          - last
                          - pages
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the request itself
                          next:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the next page of results in a paginated response.
                                  If there are no more results, this field may
                                  be omitted
                          prev:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the previous page of results in a paginated
                                  response. If there are no more results, this
                                  field may be omitted
                          first:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          last:
                            type: object
                            properties:
                              href:
                                type: string
                                description: URL pointing to the first page of results in a paginated response
                          pages:
                            type: object
                            description: It may contain a list of other useful URLs, e.g. one entry for
                              page:'page 1', 'page 2'
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        description: Object containing the links to the related endpoints
                      totalCount:
                        type: number
                        description: Number representing the total number of available items
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
                properties:
                  code:
                    type: string
                    description: Code used to categorize the error
                  detail:
                    type: string
                    description: Description of the error
                  requestId:
                    type: string
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                  name:
                    type: string
                    description: Name of the error type
                  validation:
                    type: array
                    items:
                      type: object
                      required:
                        - fieldName
                        - message
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                    description: List of the validation errors
                  validationContext:
                    type: string
                  statusCode:
                    type: number
...
security:
  - bearerAuth: []